git-svn-id: svn://svn.opensol2014.com/repo1/ezEKP4Java@2747 746f5ef6-3e3b-473a-8c3d-618c0f8083c2

+++ src/main/java/egovframework/ezEKP/ezResource/dao/EzResourceAdminDAO.java
@@ -32,7 +32,7 @@ public class EzResourceAdminDAO extends EgovAbstractDAO {
 	}
 	
 	public int delResAcll(Map<String, Object> map) {
-		return delete("EzResourceAdminDAO.delResAcll", map);
+		return (int)delete("EzResourceAdminDAO.delResAcll", map);
 	}
 	
 	public int getSubResCnt(Map<String, Object> map) {
